--- src/main/java/me/erykczy/colorfullighting/common/accessors/iris/ResolvedShaderPack.java
@@ -1,8 +1,0 @@
-package me.erykczy.colorfullighting.common.accessors.iris;
-
-import java.nio.file.Path;
-
-public interface ResolvedShaderPack {
-	String getResolvedName();
-	Path path();
-}
