+++ src/main/java/me/erykczy/colorfullighting/common/accessors/mixin/LevelAttachments.java
@@ -1,4 +1,4 @@
-package me.erykczy.colorfullighting.common.accessors;
+package me.erykczy.colorfullighting.common.accessors.mixin;
 
 import me.erykczy.colorfullighting.common.ColoredLightEngine;
