+++ src/main/java/me/erykczy/colorfullighting/common/accessors/mixin/LightEngineAccessor.java
@@ -1,4 +1,4 @@
-package me.erykczy.colorfullighting.common.accessors;
+package me.erykczy.colorfullighting.common.accessors.mixin;
 
 import net.minecraft.world.level.chunk.LightChunkGetter;
