+++ src/main/java/me/erykczy/colorfullighting/common/accessors/mixin/iris/ResolvedShaderPack.java
@@ -1,0 +1,8 @@
+package me.erykczy.colorfullighting.common.accessors.mixin.iris;
+
+import java.nio.file.Path;
+
+public interface ResolvedShaderPack {
+	String colorfullighting$getResolvedName();
+	Path colorfullighting$path();
+}
