+++ src/main/java/me/erykczy/colorfullighting/compat/oculus/CommonTransformations.java
@@ -13,7 +13,7 @@ import io.github.douira.glsl_transformer.ast.transform.ASTInjectionPoint;
 import io.github.douira.glsl_transformer.ast.transform.ASTParser;
 import io.github.douira.glsl_transformer.parser.ParseShape;
 import me.erykczy.colorfullighting.common.ColoredLightEngine;
-import me.erykczy.colorfullighting.common.accessors.iris.CustomShaderProperties;
+import me.erykczy.colorfullighting.common.accessors.mixin.iris.CustomShaderProperties;
 import me.erykczy.colorfullighting.compat.oculus.specific.ShaderSpecificPatcher;
 import me.erykczy.colorfullighting.mixin.compat.iris.ShaderPackAccessor;
 import net.irisshaders.iris.Iris;
