+++ src/main/java/me/erykczy/colorfullighting/compat/oculus/OculusCompat.java
@@ -9,7 +9,7 @@ import java.util.Optional;
 import java.util.OptionalInt;
 import java.util.zip.ZipFile;
 
-import me.erykczy.colorfullighting.common.accessors.iris.CustomShaderProperties;
+import me.erykczy.colorfullighting.common.accessors.mixin.iris.CustomShaderProperties;
 import me.erykczy.colorfullighting.mixin.compat.iris.ShaderPackAccessor;
 import net.irisshaders.iris.Iris;
 import net.irisshaders.iris.shaderpack.ShaderPack;
