+++ src/main/java/me/erykczy/colorfullighting/event/ClientEventListener.java
@@ -6,8 +6,8 @@ import me.erykczy.colorfullighting.common.BeaconEffectSync;
 import me.erykczy.colorfullighting.common.BlockEntityNbtCache;
 import me.erykczy.colorfullighting.common.ColoredLightEngine;
 import me.erykczy.colorfullighting.common.ViewArea;
-import me.erykczy.colorfullighting.common.accessors.LevelAttachments;
-import me.erykczy.colorfullighting.common.accessors.LevelRendererAccessor;
+import me.erykczy.colorfullighting.common.accessors.mixin.LevelAttachments;
+import me.erykczy.colorfullighting.common.accessors.mixin.LevelRendererAccessor;
 import me.erykczy.colorfullighting.compat.dynamiclights.DynamicLightsCompat;
 import me.erykczy.colorfullighting.compat.oculus.cmd.PackArgumentType;
 import me.erykczy.colorfullighting.compat.oculus.cmd.ShaderPackName;
