+++ src/main/java/me/erykczy/colorfullighting/mixin/BlockLightEngineMixin.java
@@ -1,8 +1,8 @@
 package me.erykczy.colorfullighting.mixin;
 
 import me.erykczy.colorfullighting.common.ColoredLightEngine;
-import me.erykczy.colorfullighting.common.accessors.LevelAttachments;
-import me.erykczy.colorfullighting.common.accessors.LightEngineAccessor;
+import me.erykczy.colorfullighting.common.accessors.mixin.LevelAttachments;
+import me.erykczy.colorfullighting.common.accessors.mixin.LightEngineAccessor;
 import net.minecraft.client.Minecraft;
 import net.minecraft.core.BlockPos;
 import net.minecraft.world.level.lighting.BlockLightEngine;
