+++ src/main/java/me/erykczy/colorfullighting/mixin/LevelChunkMixin.java
@@ -2,7 +2,7 @@ package me.erykczy.colorfullighting.mixin;
 
 import me.erykczy.colorfullighting.common.BlockEntityNbtCache;
 import me.erykczy.colorfullighting.common.ColoredLightEngine;
-import me.erykczy.colorfullighting.common.accessors.LevelAttachments;
+import me.erykczy.colorfullighting.common.accessors.mixin.LevelAttachments;
 import me.erykczy.colorfullighting.common.util.ShapeOcclusion;
 import net.minecraft.client.Minecraft;
 import net.minecraft.core.BlockPos;
